+++ scripts/character.gd
@@ -1,88 +1,156 @@
 extends CharacterBody2D
+class_name Character
+## The PNG of this character's sprite sheet
+@export var SpriteSheet:CompressedTexture2D
 
-# The proper name of the character
+
+
+@export_group("Character Source")
+## The proper name of the character
 @export var CharacterName:String
+## The name of the game or media this character came from
+@export var CharacterSource:String
+enum characterSourceTypes {Game,Interpretation,RelatedGame,RomHack,FanGame,Original}
+## The type of media that the source is
+@export var CharacterSourceType:characterSourceTypes
+## The name of the original artist that created this sprite (or its source), if known
+@export var OriginalArtist:String
 
-var controller : Node
 
-const SPEED = 100.0
-const FLAP_VELOCITY = -100.0
-const JUMP_VELOCITY = -50.0
-const GRAVITY_MULTIPLIER = 0.2
-const AIR_FRICTION = 1
+@export_group("Character Abilities")
+## Whether this character should be able to use a parachute while falling
+@export var HasParachute:bool = false
 
+
+@export_group("Character Stats")
+# How quickly the player comes to a stop after letting go of run button (0.0-0.1)
+@export var LandFriction:float = 0.2
+# The fastest this character can travel while in the air
+@export var TopAirSpeed:float = 100
+# The fastest this character can travel while on the ground
+@export var TopLandSpeed:float = 100
+# The vertical speed attained when this character jumps off the ground
+@export var JumpHeight:float = 50
+# The amount of gravity the character expiriences when they have at least 1 balloon (multiplied with the normal gravity)
+@export var BalloonGravity:float = 0.4
+# The percentage of their full TopAirSpeed that the character reaches with each flap
+@export var FlapStrength:float = 0.2
+
+# set by code on setup
+var controller:Node
+var playerHUD:PlayerHUD
 var playerNumber:int
-
 var gravity = ProjectSettings.get_setting("physics/2d/default_gravity")
 
-var current_balloons = 4
+# player state variables
+var current_balloons = 0
+var stored_balloons = 0
 var falling = false
+var parachuting = false
 var invincible = false
-
-func init():
-	print("READY")
-	$Balloons.animation = str(current_balloons) + "_idle"
-
-func _draw():
-	if controller:
-		if controller.has_method("debugDraw"):
-			controller.debugDraw(self)
-		if "DEBUG" in controller and controller.DEBUG and invincible:
-			draw_circle(Vector2(-15,-15),1,Color(0,1,1))
+var current_fill_amount = 0
+	
 
 func _physics_process(delta):
-	if controller and controller.has_method("debugDraw"): queue_redraw()
-	
+	if DEBUG or (controller and controller.has_method("debugDraw")): queue_redraw()
 	if invincible && Time.get_ticks_msec() > invincible: invincible = false
 	
-	# Add the gravity.
+	# Falling
 	if falling:
 		velocity.y += gravity * delta
-		if position.y > 360: queue_free()
+		if position.y > 400: queue_free()
 		return move_and_slide()
+	# Falling with parachute
+	elif parachuting:
+		if is_on_floor():
+			parachuting = false
+			$Balloons.animation = "0_idle"
+		else:
+			velocity.x = lerp(velocity.x,sin(Time.get_ticks_msec() / 500 ) * 32, 0.1)
+			velocity.y += gravity*BalloonGravity*0.25 * delta
+			if position.y > 360: queue_free()
+			return move_and_slide()
+	# Jumping without balloon
+	elif not is_on_floor() and current_balloons == 0:
+		velocity.y += gravity * delta
+	# Floating with balloon
 	elif not is_on_floor():
-		velocity.y += gravity*GRAVITY_MULTIPLIER * delta
+		velocity.y += gravity*BalloonGravity * delta
 	
 	if not controller: return move_and_slide()
 	var input = controller.getInput(self)
 	
-	# Handle Jumping
+	# Handle Jumping / Flapping
 	if input.jumped:
-		$Sprite.animation = 'flap'
-		$Sprite.play()
-		if is_on_floor(): velocity.y = JUMP_VELOCITY
-		else: velocity.y = lerp(velocity.y, FLAP_VELOCITY, 0.15)
+		if is_on_floor(): 
+			if current_balloons > 0:
+				velocity.y = -JumpHeight
+			else:
+				velocity.y = (-JumpHeight) * 2
+		elif current_balloons > 0: 
+			$Sprite.animation = 'flap'
+			$Sprite.play()
+			velocity.y = lerp(velocity.y, -TopAirSpeed, FlapStrength)
 	
+	# Balloon Filling
+	var filling=false
+	if $Sprite.animation == 'refill': filling = true
+	if input.fill and is_on_floor() and velocity.x < 5:
+		if $Sprite.animation != 'refill' or !$Sprite.is_playing():
+			if stored_balloons > 0 and current_balloons < 4:
+				current_fill_amount+=1
+				filling = true
+				$Audio/Fill.play()
+				$Sprite.play("refill")
+				if current_fill_amount>2: $BalloonFilling.play("fill_2")
+				else: $BalloonFilling.play("fill_1")
+			else: # TODO play nono sound effect
+				pass
+	if input.dirAxis or input.jumped:
+		current_fill_amount = 0
+		filling=false
+		$BalloonFilling.play("blank")
+		
 	# Handle Movement
 	if input.dirAxis:
-		if is_on_floor():
-			velocity.x = input.dirAxis * SPEED
-		else:
-			velocity.x = move_toward(velocity.x, input.dirAxis * SPEED, AIR_FRICTION)
 		$Sprite.flip_h = input.dirAxis > 0
-	else:
 		if is_on_floor():
-			velocity.x = move_toward(velocity.x, 0, SPEED)
-		else:
-			velocity.x = move_toward(velocity.x, 0, 1)
-	
+			velocity.x = input.dirAxis * TopLandSpeed
+		elif input.jumped:
+			velocity.x = lerp(velocity.x, input.dirAxis * TopAirSpeed, 0.2)
+	elif is_on_floor():
+		velocity.x = lerp(velocity.x, 0.0, LandFriction)
+
+
 	# Play Animations
-	if (is_on_floor()):
-		if (velocity.x == 0): $Sprite.animation = 'idle'
+	if is_on_floor():
+		if filling: $Sprite.animation = 'refill'
+		elif abs(velocity.x) < 25: $Sprite.animation = 'idle'
 		else: $Sprite.animation = 'walk'
 	else:
 		if ($Sprite.animation != 'flap'): $Sprite.animation = 'float'
 	
 	# Move
 	move_and_slide()
-	
+
 	# Check Collisions
 	for i in get_slide_collision_count():
 		var collision = get_slide_collision(i)
+		var localShape = collision.get_local_shape()
+		var otherShape = collision.get_collider_shape()
+		#print("shape ",	localShape, " disabled:",localShape.disabled)
+		if localShape.disabled or otherShape and otherShape.disabled: continue
 		#print("I collided with ", collision.get_collider().name)
-		
-		if collision.get_collider().is_in_group("character"):
-			print("\n[",i," collision with ",collision.get_collider().name,"] normal: ",collision.get_normal(), " | dot: ", Vector2.UP.dot(collision.get_normal()))
+		var collider = collision.get_collider()
+		if collider.is_in_group("character"):
+			if (falling): break
+			if current_balloons == 0 and !falling:
+				startFalling()
+				break
+			if collider.current_balloons == 0 and !collider.falling:
+				collider.startFalling()
+			var normal = collision.get_normal()
+			#print("\n[",i," collision with ",collider.name,"] normal: ",normal, " | dot: ", Vector2.UP.dot(normal))
 			
 			var travel = collision.get_travel() 
 			var remainder = collision.get_remainder()
@@ -98,17 +166,41 @@ func loseBalloon ():
 	print("POOOPPPPPPPPPPPPPPPPPPPPPPPPPP",name)
 	$Balloons.animation = str(current_balloons) + "_pop"
 	current_balloons -= 1
+	if current_balloons < 0: printerr("current_balloons dropped below 0, should be impossible", name, current_balloons)
 	$Audio/Pop.playing = true
 	if current_balloons == 0:
-		falling = true
-		collision_mask = 0
-		$Sprite.animation = "fall"
+		switchCollisionShape()
+		if HasParachute:
+			$Sprite.animation = "parachute"
+			$Balloons.animation = "parachute"
+			parachuting = true
+			invincible = Time.get_ticks_msec() + 500
+		else:
+			startFalling()
 	else:
 		invincible = Time.get_ticks_msec() + 1000
-	
+
+func startFalling():
+	if (falling): return
+	print("start falling...")
+	falling = true
+	collision_mask = 0
+	$Sprite.animation = "fall"
+	$Audio/Falling.play(0)
+	playerHUD.update()
 
 func _on_sprite_animation_finished():
-	if $Sprite.animation == 'flap':	$Sprite.animation = 'float'
+	if $Sprite.animation == 'flap':	
+		$Sprite.animation = 'float'
+	if $Sprite.animation == 'refill':
+		if stored_balloons > 0 and current_fill_amount>=4:
+			current_fill_amount = 0
+			stored_balloons -= 1
+			current_balloons += 1
+			playerHUD.update()
+			$BalloonFilling.play("blank")
+			$Balloons.animation = str(current_balloons) + "_idle"
+			$Audio/Filled.play()
 
 func _on_sprite_animation_changed():
 	$Sprite.play()
@@ -118,14 +210,98 @@ func _on_balloons_animation_changed():
 	
 func _on_balloons_animation_finished():
 	print("finito",$Balloons.animation)
-	if current_balloons && $Balloons.animation.ends_with("pop"):
+	if $Balloons.animation.ends_with("pop"):
 		$Balloons.animation = str(current_balloons) + "_idle"
+
 
 func _on_balloon_area_body_shape_entered(_body_rid, body, _body_shape_index, _local_shape_index):
 	print("balloon shape entered", body.name)
 	
 func _on_balloon_area_area_entered(area):
-	if not invincible:
-		if area.name == "FeetArea": 
-			print("balloon area entered", area.name)
-			loseBalloon()
+	if invincible: return
+	if current_balloons == 0: return
+	var otherCharacter = area.get_parent()
+	if otherCharacter.falling or otherCharacter.parachuting or otherCharacter.current_balloons < 1: return
+	if area.name == "FeetArea": 
+		print("balloon area entered", area.name)
+		
+		#position.angle_to(otherCharacter.position)
+		var collision_angle = position.angle_to(otherCharacter.position)
+		# Calculate the bounce angles for each character
+		var bounce_angle1 = collision_angle + PI / 2  # 90 degrees offset
+		var bounce_angle2 = collision_angle - PI / 2  # -90 degrees offset
+
+		# Set the velocities based on the bounce angles and desired speed
+		velocity = Vector2(cos(bounce_angle1), sin(bounce_angle1)) * 40
+		otherCharacter.velocity = Vector2(cos(bounce_angle2), sin(bounce_angle2)) * 75
+		loseBalloon()
+
+func switchCollisionShape():
+	print("switing collision shape",current_balloons == 0)
+	if current_balloons == 0:
+		$HasBallonsCollisionShape.set_deferred("disabled", true)
+		$NoBallonsCollisionShape.set_deferred("disabled", false)
+	else:
+		$HasBallonsCollisionShape.set_deferred("disabled", false)
+		$NoBallonsCollisionShape.set_deferred("disabled", true)
+	
+	print("$HasBallonsCollisionShape.disabled",$HasBallonsCollisionShape.disabled)
+	print("$NoBallonsCollisionShape.disabled",$NoBallonsCollisionShape.disabled)
+
+
+func init():
+	$Balloons.animation = str(current_balloons) + "_idle"
+	update_texture($Sprite, SpriteSheet)
+	update_texture($Balloons, SpriteSheet)
+
+# changes a sprite texture at the beginning of the match
+func update_texture(sprite: AnimatedSprite2D, texture: Texture2D):
+	var reference_frames := sprite.sprite_frames
+	var updated_frames := SpriteFrames.new()
+
+	for animation_name in reference_frames.get_animation_names():
+		if animation_name != "default":
+			updated_frames.add_animation(animation_name)
+			updated_frames.set_animation_speed(animation_name, reference_frames.get_animation_speed(animation_name))
+			updated_frames.set_animation_loop(animation_name, reference_frames.get_animation_loop(animation_name))
+
+			for i in range(reference_frames.get_frame_count(animation_name)):
+				var original_frame := reference_frames.get_frame_texture(animation_name, i) as AtlasTexture
+				var updated_texture := original_frame.duplicate() as AtlasTexture
+				updated_texture.atlas = texture
+
+				# Copy the region from the original frame
+				updated_texture.region = original_frame.region
+
+				# Find the index of the frame with the matching texture
+				var frame_index := -1
+				for j in range(updated_frames.get_frame_count(animation_name)):
+					if updated_frames.get_frame_texture(animation_name, j) == original_frame:
+						frame_index = j
+						break
+
+				# If the frame is found, update it
+				if frame_index != -1:
+					updated_frames.set_frame(animation_name, frame_index, updated_texture)
+				else:
+					# Add the frame with the correct duration if not found
+					updated_frames.add_frame(animation_name, updated_texture, reference_frames.get_frame_duration(animation_name, i))
+
+	updated_frames.remove_animation("default")
+	sprite.sprite_frames = updated_frames
+	sprite.play() # Ensure the animation plays
+	
+const DEBUG = false
+@onready var default_font = ThemeDB.fallback_font
+func _draw():
+	if controller:
+		if controller.has_method("debugDraw"):
+			controller.debugDraw(self)
+	if DEBUG:
+		if invincible: draw_circle(Vector2(-15,-15),1,Color(0,1,1))
+		draw_string(default_font, Vector2(10, -5), str(current_balloons), HORIZONTAL_ALIGNMENT_LEFT, -1, 8, Color.WHITE)
+		if !$HasBallonsCollisionShape.disabled: draw_string(default_font, Vector2(10, 5),'c:has', HORIZONTAL_ALIGNMENT_LEFT, -1, 8, Color.WHITE)
+		if !$NoBallonsCollisionShape.disabled: draw_string(default_font, Vector2(10, 5),'c:no', HORIZONTAL_ALIGNMENT_LEFT, -1, 8, Color.WHITE)
+		if current_fill_amount > 0:
+			draw_string(default_font, Vector2(0, -15), str(current_fill_amount), HORIZONTAL_ALIGNMENT_LEFT, -1, 8, Color.WHITE)
+